# Minor perf improvements