**This PR:**
- separated tests from benchmarks
- added gitignore
- made some minor perf improvements.

```
benchmark             old ns/op     new ns/op     delta
BenchmarkSmall        2452          2245          -8.44%
BenchmarkSmall-2      2430          2237          -7.94%
BenchmarkSmall-4      2454          2293          -6.56%
BenchmarkMedium       4490          4337          -3.41%
BenchmarkMedium-2     4537          4559          +0.48%
BenchmarkMedium-4     4589          4289          -6.54%
BenchmarkLarge        19574         17468         -10.76%
BenchmarkLarge-2      19470         17889         -8.12%
BenchmarkLarge-4      20005         18347         -8.29%

benchmark             old allocs     new allocs     delta
BenchmarkSmall        17             17             +0.00%
BenchmarkSmall-2      17             17             +0.00%
BenchmarkSmall-4      17             17             +0.00%
BenchmarkMedium       30             30             +0.00%
BenchmarkMedium-2     30             30             +0.00%
BenchmarkMedium-4     30             30             +0.00%
BenchmarkLarge        118            107            -9.32%
BenchmarkLarge-2      118            107            -9.32%
BenchmarkLarge-4      118            107            -9.32%

benchmark             old bytes     new bytes     delta
BenchmarkSmall        1188          1188          +0.00%
BenchmarkSmall-2      1188          1188          +0.00%
BenchmarkSmall-4      1189          1188          -0.08%
BenchmarkMedium       2707          2706          -0.04%
BenchmarkMedium-2     2707          2707          +0.00%
BenchmarkMedium-4     2707          2707          +0.00%
BenchmarkLarge        13856         13327         -3.82%
BenchmarkLarge-2      13856         13328         -3.81%
BenchmarkLarge-4      13856         13328         -3.81
```